Insight/requirements.txt
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.5
django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.9
openapi-codec==1.3.1
pkg-resources==0.0.0
requests==2.13.0
simplejson==3.10.0
uritemplate==3.0.0
